Move http handlers from styx-api to styx-common module. This cuts down Google Guava dependencies in the API module.
