+++ components/common/src/main/java/com/hotels/styx/common/http/handler/StaticBodyHttpHandler.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.http.handlers;
+package com.hotels.styx.common.http.handler;
 
 import com.google.common.base.Charsets;
 import com.google.common.net.MediaType;
