+++ components/common/src/test/java/com/hotels/styx/common/http/handler/StaticBodyHttpHandlerTest.java
@@ -13,16 +13,16 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.http.handlers;
+package com.hotels.styx.common.http.handler;
 
 
 import com.hotels.styx.api.FullHttpResponse;
 import com.hotels.styx.api.HttpResponse;
+import com.hotels.styx.server.HttpInterceptorContext;
 import org.testng.annotations.Test;
 
 import static com.google.common.net.MediaType.PLAIN_TEXT_UTF_8;
 import static com.hotels.styx.api.HttpRequest.get;
-import static com.hotels.styx.api.MockContext.MOCK_CONTEXT;
 import static com.hotels.styx.api.messages.HttpResponseStatus.OK;
 import static com.hotels.styx.support.matchers.IsOptional.isValue;
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -34,7 +34,7 @@ public class StaticBodyHttpHandlerTest {
     public void respondsWithStaticBody() throws Exception {
         StaticBodyHttpHandler handler = new StaticBodyHttpHandler(PLAIN_TEXT_UTF_8, "foo", UTF_8);
 
-        HttpResponse response = handler.handle(get("/").build(), MOCK_CONTEXT).asCompletableFuture().get();
+        HttpResponse response = handler.handle(get("/").build(), HttpInterceptorContext.create()).asCompletableFuture().get();
         FullHttpResponse fullResponse = response.toFullResponse(1024)
                 .asCompletableFuture()
                 .get();
