+++ components/proxy/src/main/java/com/hotels/styx/admin/AdminServerBuilder.java
@@ -42,8 +42,8 @@ import com.hotels.styx.admin.tasks.OriginsCommandHandler;
 import com.hotels.styx.admin.tasks.OriginsReloadCommandHandler;
 import com.hotels.styx.api.HttpHandler;
 import com.hotels.styx.api.configuration.Configuration;
-import com.hotels.styx.api.http.handlers.HttpMethodFilteringHandler;
-import com.hotels.styx.api.http.handlers.StaticBodyHttpHandler;
+import com.hotels.styx.common.http.handler.HttpMethodFilteringHandler;
+import com.hotels.styx.common.http.handler.StaticBodyHttpHandler;
 import com.hotels.styx.api.service.BackendService;
 import com.hotels.styx.api.service.spi.Registry;
 import com.hotels.styx.proxy.plugin.NamedPlugin;
