+++ components/proxy/src/main/java/com/hotels/styx/admin/handlers/IndexHandler.java
@@ -17,7 +17,7 @@ package com.hotels.styx.admin.handlers;
 
 import com.hotels.styx.api.FullHttpResponse;
 import com.hotels.styx.api.HttpResponse;
-import com.hotels.styx.api.http.handlers.BaseHttpHandler;
+import com.hotels.styx.common.http.handler.BaseHttpHandler;
 import com.hotels.styx.api.HttpRequest;
 
 import static com.google.common.base.Preconditions.checkNotNull;
