+++ components/proxy/src/main/java/com/hotels/styx/admin/handlers/ThreadsHandler.java
@@ -19,7 +19,7 @@ import com.codahale.metrics.jvm.ThreadDump;
 import com.hotels.styx.api.FullHttpResponse;
 import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.api.HttpResponse;
-import com.hotels.styx.api.http.handlers.BaseHttpHandler;
+import com.hotels.styx.common.http.handler.BaseHttpHandler;
 
 import java.io.ByteArrayOutputStream;
