+++ components/proxy/src/test/java/com/hotels/styx/admin/handlers/StatusHandlerTest.java
@@ -18,7 +18,7 @@ package com.hotels.styx.admin.handlers;
 import com.hotels.styx.api.FullHttpResponse;
 import com.hotels.styx.api.HttpHandler;
 import com.hotels.styx.api.HttpResponse;
-import com.hotels.styx.api.http.handlers.BaseHttpHandler;
+import com.hotels.styx.common.http.handler.BaseHttpHandler;
 import com.hotels.styx.api.messages.HttpResponseStatus;
 import com.hotels.styx.server.HttpInterceptorContext;
 import org.testng.annotations.Test;
