+++ components/server/src/main/java/com/hotels/styx/server/handlers/ClassPathResourceHandler.java
@@ -20,7 +20,7 @@ import com.hotels.styx.api.FullHttpResponse;
 import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.api.HttpResponse;
 import com.hotels.styx.api.StyxObservable;
-import com.hotels.styx.api.http.handlers.BaseHttpHandler;
+import com.hotels.styx.common.http.handler.BaseHttpHandler;
 import com.hotels.styx.api.messages.HttpResponseStatus;
 
 import java.io.FileNotFoundException;
