+++ components/server/src/main/java/com/hotels/styx/server/handlers/StaticFileHandler.java
@@ -32,7 +32,7 @@ import java.util.Optional;
 import static com.google.common.base.Throwables.propagate;
 import static com.hotels.styx.api.HttpHeaderNames.CONTENT_TYPE;
 import static com.hotels.styx.api.MediaTypes.mediaTypeOf;
-import static com.hotels.styx.api.http.handlers.NotFoundHandler.NOT_FOUND_HANDLER;
+import static com.hotels.styx.common.http.handler.NotFoundHandler.NOT_FOUND_HANDLER;
 import static com.hotels.styx.api.messages.HttpResponseStatus.INTERNAL_SERVER_ERROR;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.slf4j.LoggerFactory.getLogger;
