+++ support/origins-starter-app/src/main/java/com/hotels/styx/support/origins/AppHandler.java
@@ -21,7 +21,7 @@ import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.api.HttpResponse;
 import com.hotels.styx.api.StyxObservable;
 import com.hotels.styx.api.client.Origin;
-import com.hotels.styx.api.http.handlers.StaticBodyHttpHandler;
+import com.hotels.styx.common.http.handler.StaticBodyHttpHandler;
 import com.hotels.styx.api.messages.HttpResponseStatus;
 
 import static com.google.common.net.MediaType.HTML_UTF_8;
