+++ system-tests/e2e-suite/src/test/scala/com/hotels/styx/MockServer.scala
@@ -22,8 +22,8 @@ import java.util.function.Supplier
 import com.google.common.util.concurrent.AbstractIdleService
 import com.hotels.styx.api._
 import com.hotels.styx.api.client.Origin.newOriginBuilder
-import com.hotels.styx.api.http.handlers.NotFoundHandler
 import com.hotels.styx.api.support.HostAndPorts._
+import com.hotels.styx.common.http.handler.NotFoundHandler
 import com.hotels.styx.server.handlers.ReturnResponseHandler.returnsResponse
 import com.hotels.styx.server.netty.{NettyServerBuilder, ServerConnector, WebServerConnectorFactory}
 import com.hotels.styx.server.{HttpConnectorConfig, HttpServer}
